tests/node_sim/requirements.txt
# Node Simulator - зависимости
paho-mqtt>=1.6.1
pyyaml>=6.0
httpx>=0.24.0
websocket-client>=1.6.0
psycopg[binary]>=3.1.0
tenacity>=8.2.0
cachetools>=5.3.0

